Makefile: Build/install a static library in addition to the dylib.
New convenience sub-targets: `slib`, `dlib`, `install-slib`, `install-dlib`.

It is possible to add specific CFLAGS for the static resp. dynamic lib.
This is now used to compile with -fPIC for the dylib (although -fPIC is
the default on Darwin anyway). Dylib object files are named *.dl.o now.

Note: test programs are linked against the dylib.
To test the static library, you could run the following command:

    make clean all check TESTLIBS=lib/lib*.a