Use asm symbol declarations for wrappers
The idea is taken from Apple's own declaration of `realpath`:
It boils down to

    extern char *realpath(const char * __restrict, char * __restrict)
                 asm("_linker_symbol_for_realpath");

This means that we do not need to use macros for renaming symbols
(except for temporarily moving the original declaration out of the way).
The compiler does that for us where necessary (and only there).
This preserves source-level API compatibility.
Take a look at test/test_{realpath,sysconf}.c for the scenarios tested.