add Tiger-specific fixes
initial libproc support for Tiger
still missing proc_pidpath()
but these headers provide access to the functions in libSystem

a slightly modified 'which' command that returns nothing
if the requested tool is not found, rather than returning a
message indicating that nothing was found. Returning nothing
is the current behaviour of 'which' and a number of scripts
rely on this behaviour.