Makefile: add FORCE_ARCH variable.
Can be used to specify an architecture directly and bypass the lipo
-archs check, that might require a newer version of cctools than the
base system/Xcode/CLTs provide.

Mostly only useful for MacPorts.

Note that forcing an architecture also requires the compiler to generate
single-architecture files. This is not done for technical reasons, since
it could equally well be an architecture list and used as such, but for
convenience and a safety check when using the muniversal PortGroup
within MacPorts.