utimensat: test fixups.
Add darwintest emulation macros (as far as we need them) and debugging
messages to see where things fail.

This test is known to work on 10.13 + APFS (which uses the native OS
implementation of utimensat), but known to fail on HFS+, regardless of
the utimensat implementation.

Hence, it needs to be rewritten to test for microsecond accuracy on
non-APFS file systems.