setattrlistat: emulate 10.8 behavior in older systems.
Starting with 10.8, the internal setattrlist code checks if no
attributes to set have been given and exits early, fixing a bug where an
attribute buffer size of zero leads to it exiting with an error (which,
really, should be mostly a no-op).

This should fix the utimensat test case on 10.7 and below.