add support for MACOSX_DEPLOYMENT_TARGET
rust-bootstrap runs on a system that is older than the one on which
it was built.
So `rustc` from rust-bootstrap needs to know to look in
libMacportsLegacySystem.B.dylib for symbols even if they are provided
by the system on which it was built.

Use the `__asm("$ld$add$os10.x$...")` trick, which respects
`MACOSX_DEPLOYMENT_TARGET`.