time.c: Fix broken CLOCK_MONOTONIC
Commit bb525c1 replaced the subtly incorrect CLOCK_MONOTONIC fallback
with a version that isn't even monotonic.  This replaces it with a
version that uses the mach SYSTEM_CLOCK, which has been in use in
ntpsec since Jan-2019.

TESTED:
Ran "port destroot" and "port test" on legacy-support-devel, modified
to reference this commit.  Ran on 10.5 ppc, 10.5-10.6 i386, 10.5-10.15
x86_64, and 11.x-14.x arm64.  Built in all cases, and passed tests in
all cases except previously failing ones.
Also ran on 10.4 ppc and 10.4 i386, but using this same patch as
applied to the 1.0.13 release, since the current master doesn't build
on 10.4.  Built and passed tests except for ppc +universal, which
fails to build as before.