test_fdopendir: Add pipe tests
The stdin test was failing (legacy version only) when the test was run
from a pipe rather than a tty or pty.  This adds an explicit test for
the pipe case, allowing that to be tested even while the program is
being run from a tty or pty.  The existing stdin test can also be
observed to fail when running the test via "ssh -T" or by running:

	 cat /dev/null | .../test_fdopendir

Since a pipe is now created for this test, it also tests the
closed-pipe case, as a general test of a closed-fd case, which isn't
necessarily the same as a numerically invalid fd (though it should
return the same EBADF).

TESTED:
Without new fdopendir() fixes, the open-pipe test fails in all <10.10
OSes, and the closed-pipe test fails in 10.4.  With new fixes, all
tests pass in all OS versions.
Tested on 10.4-10.5 ppc, 10.4-10.6 i386, 10.4-12.x x86_64, 11.x-14.x
arm64.