fdopendir: Clean up and fix OS version check
This defines the 10.5+ __dd_fd as a macro for the 10.4 dd_fd, thereby
aviding the need for a conditional in the body of the code.

Also, exact equality comparisons for ..._MIN_REQUIRED__ are
inappropriate.  In particular, in 10.4,
__ENVIRONMENT_MAC_OS_X_VERSION_MIN_REQUIRED__ may not be defined at
all, though the assumed default value of 0 has the right effect on
inequality comparisons.  Also, it's best to use "< <next os>" rather
than "<= <this os>", to allow for multiple values of the low-order
digit.

TESTED:
Still works for all OS versions.
Tested on 10.4-10.5 ppc, 10.4-10.6 i386, 10.4-12.x x86_64, 11.x-14.x
arm64.