sysconf(_SC_PHYS_PAGES): Always use wrapper on i386
Apple's implementation of sysconf(_SC_PHYS_PAGES) never worked on
i386.  It wasn't provided until 10.11, which requires x86_64, so the
problem is never seen in ordinary native builds.  It is visible in
i386 builds for 10.11+, including universal builds with i386.  Since
MacPorts dropped i386 from the default universal_archs in 10.14, it's
not normally seen in 10.14+.  It can be seen in the test failures in
the 10.11-10.13 +universal cases.

The fix is simply to use the working legacy-support implementation in
all i386 builds, regardless of OS version.

TESTED:
Tested on 10.4-10.5 ppc, 10.5-10.6 ppc (x86_64 Rosetta), 10.4-10.6
i386, 10.4-12.x x86_64, and 11.x-14.x arm64.
Builds on all tested platforms except 10.4 ppc +universal, and passes
all tests in all buildable cases, including the previously failing
10.11-10.13 +universal tests.