Add optional security wrapper for stpncpy().
This adds support for the optional stpncpy() security wrapper, based
on the _FORTIFY_SOURCE setting.  By default, it only impacts 10.6
builds, though it can be explicitly enabled on 10.5, albeit less
efficiently due to the lack of the compiler builtin.

The stpncpy() function is the only one with an optional security
wrapper which is also optionally provided by legacy-support.  Hence,
this is the only addition needed to close the more general ticket.

Closes: https://trac.macports.org/ticket/69878

Also fixes a minor comment formatting issue.

TESTED:
Tested on 10.4-10.5 ppc, 10.4-10.6 i386, 10.5-10.6 ppc (i386 Rosetta),
10.5-12.x x86_64, 11.x-14.x arm64.
Passes all tests, including newly added tests for this feature.