Move renameat() to the proper header.
Commit e9cf5ab3 moved the renameat() declaration from sys/stdio.h to
stdio.h, but this is incompatible with the normal header setup.
This moves it back to restore compatibility, but requires knowledge
of the SDK version to cope with the absence of sys/stdio.h in
pre-10.10 SDKs.  This is handled via the new sdkversion.h.

Re: https://trac.macports.org/ticket/69867

TESTED:
Builds and passes tests on all platforms, including the new test for
this issue.