Add sys/cdefs.h wrapper, and tests.
This is needed to ensure that the __DARWIN_C_* definitions are
available with SDKs prior to 10.7.

Also adds a test for the effects of __DARWIN_C_LEVEL, initially just
as a test that the three macros are defined.

Also adds a new manual test, which simply reports the values of the
relevant macros.

Re: https://trac.macports.org/ticket/69688

TESTED:
Test fails without the new sys/cdefs.h, and passes with it.
Manual test reports as expected.