Add tests for strncpy_chk(), for comparison.
This is done purely for comparison purposes, by testing security
wrappers for strncpy(), which is *not* provided by legacy-support.
Since the functions are identical except for the return value, which
is ignored by these tests, this is a near-verbatim copy of the
strncpy_chk tests, except for the name.

TESTED:
Behaves as expected, including failing some 10.4 with 10.5+ SDK cases,
prior to the fix.