Tighten test_stpncpy_chk.
Instead of disabling the tests in the no-check case, it now explicitly
tests for non-failure of the failure tests in that case, as well as
unconditionally checking for non-failure of the non-failure case, but
only in the "static" case, since the "auto" check seems to happen even
when the checks are nominally disabled.

This doesn't apply to the disabled runtime check tests.

Also adds a check for the explicitly disabled case.

TESTED:
Passes normally in all cases.
Now fails when tweaked to treat an enabled case as a disabled case.