Fix stpncpy() security wrapper for 10.5 with 10.7+ SDK.
The previous fix for the 10.5 compiler issue only applied when the
wrapper is being provided by the legacy-support headers.  But the
problem also exists when the wrapper is applied by a 10.7+ SDK when
building for 10.5.  This moves the hack outside the check for the
prefdefined wrapper.

In the 10.7+ SDK case, the inline was defined before the macro hack,
so we use a different name for the inline here.  To make this
effective, we unconditionally define the macro here, overriding the
SDK.

Re: https://trac.macports.org/ticket/69867

TESTED:
Tested on 10.4-10.5 ppc, 10.5-10.6 ppc (i386 Rosetta), 10.4-10.6 i386,
10.5-12.x x86_64, and 11.x-14.x arm64.  Tested against all 10.4-14.x
SDKs, using the headerinfo manual test (from a subsequent commit).
All target/SDK combinations where the SDK supports the CPU
architecture work correctly, including all SDKs on Intel
architectures.  Only 10.4-10.6 SDKs support ppc, and only 11.x+ SDKs
support arm64.