Makefile: Fix CC/CXX/LD flags issues.
This fixes three issues with specifying flag variables to the Makefile:

1) It attempted to insert ARCHFLAGS into CC and CXX, but it was using ?=,
which doesn't work for predefined variables.

2) It never incorporated ARCHFLAGS into LDFLAGS.

3) It provided some defaults in CFLAGS and CXXFLAGS, but any setting of
these variables removed those defaults.  The defaults are now in XCFLAGS
and XCXXFLAGS, which are added to C[XX]FLAGS, along with ARCHFLAGS, so
they normally are kept.  If removing the defaults is desired, then
XC[XX]FLAGS can be overridden.

Note that the port build procedure uses FORCE_ARCH, not ARCHFLAGS.

Closes: https://trac.macports.org/ticket/69782

TESTED:
ARCHFLAGS is now respected, and setting CFLAGS doesn't remove
defaults.