Fix CLOCK_UPTIME_RAW_APPROX definition.
This constant was misdefined to be the same value as CLOCK_UPTIME_RAW.
In the simple case, it made no difference since both cases are treated
identically by the library.  But that meant that the library didn't
support CLOCK_UPTIME_RAW_APPROX with the correct value, which could
occur if the client were built with an SDK that defined it.  This can
be seen by running test_time.cpp built with an SDK>=10.12 on OS<10.12,
though only in the text output, since the test "succeeds" even when
the functions don't.

TESTED:
The aforementioned failing case now works properly.