fmemopen: Add missing conditional.
The implementation of fmemopen() was lacking the usual conditional, so
that it was present even in OS versions where the OS provides the
function.  At best, this was a bit of code bloat.  At worst, it might
substitute an inferior implementation in later OS versions, rather
than getting out of the way.

TESTED:
Tests pass.
The fmemopen() function is now absent on 10.13 builds, while still
being present in 10.12 builds.