Use proper SDK/lib conditions for stpncpy().
This splits the feature flag for stpncpy() into separate flags for the
target OS version and the SDK version (the first of many such cases).
For the library, this just renames the feature flag.  For the header,
it's complicated by the added feature of fixing the possible
incompatibility between 10.7+ SDKs and old compilers.  See the
comments in string.h.

TESTED:
Tests pass on all platforms, including mismatched SDK cases.