Remove superfluous ifndefs from some headers.
Some constant definitions were wrapped in #ifndef/#endif to defend
against the possibility of building with a later SDK that provided
those definitions.  But in cases where a proper SDK version
conditional is used, this is redundant.

TESTED:
Builds and passes tests on all platforms, including with allowable
mismatched SDK cases.