Rename MACPORTS_LEGACY_MIN_EARLY_SDK_ALLOWED
It's more consistent with normal naming conventions to use a
leading underscore for the name of this flag.  Thus, the
flag is now _MACPORTS_LEGACY_MIN_EARLY_SDK_ALLOWED.

Since this feature is only recently added and rarely needed, it's
not too likely that this change will break anything.

TESTED:
Updated name has desired effect.