Add scandir() compatibility feature.
This adds a wrapper to make the scandir() 'compar' argument and the
provided alphasort() function match the types used in 10.8+, thereby
avoiding warnings, and in some cases errors, due to the pointer type
mismatch.

To avoid breaking code expecting the old behavior, this hack is only
applied when _MACPORTS_LEGACY_COMPATIBLE_SCANDIR is defined nonzero.

Closes: https://trac.macports.org/ticket/70702

TESTED:
Tests pass on all OS versions with this fix.