Fix some builds (e.g. Xcode 7.2 clang 7) with 15.x SDK.
Due to a clang bug, the 15.x SDK sys/cdefs.h failed to compile with
the XCode 7.2 clang 7 on OS 10.10.  This adds a workaround for that.

This bug is also present in some gcc versions, so the workaround is
applied there as well, albeit with an apparently unavoidable warning.

TESTED:
The checksdkversion and darwin_c_all tests now build and pass with the
15.x SDK and the usual Xcode compiler on all OS versions, as well as
all non-problematic MacPorts compilers tested.