Deprecate C++ dirent test.
The C++ test for dirent.h was intended to test a former problem that
no longer exists in the new fdopendir() implementation.  Since C++
tests have become somewhat problematic, that test has been moved to
the manual tests.  Given that it's now rather pointless, it's been
replaced with a minimal test of just the include, rather than
rewriting the whole test in C.

TESTED:
Passes on all OS versions with all SDKs.