Fix SDK 15 math.h for compilers without _Float16
Although Apple usually provides workarounds for newer SDKs with older
compilers, they neglected to do that for the prototypes of the new
half-precision floating-point functions added in macOS 15.0.  With
older compilers, this results in failed builds, even for code that
doesn't use the new functions.  Here we provide a workaround for that.

Also makes the _Float16 support status available as a macro,
regardless of SDK version.

TESTED:
Xcode compilers for all OS versions, as well as many others, now
successfully build math.h from all SDKs.