Rewrite test_time in C.
This avoids some C++ problems introduced with the 15.x SDK.  The
original C++ test is retained as a manual test.

This version also treats failure returns from the calls being tested
as errors.  The C++ version simply output the failing return values to
the terminal and then "succeeded".

The lack of such checks in the old version covered up the fact that
CLOCK_MONOTONIC was broken on 10.4.  This is now fixed in the previous
commit.

TESTED:
Passes on all OS versions with all "non-earlier" SDKs.