Fix CLOCK_MONOTONIC for 10.4
Referencing "kern.boottime" via sysctlbyname() on 10.4 doesn't work,
though it works to use the numeric constants with sysctl().  It was
always really dumb to be doing a string lookup on every call, anyway,
and this is a speed improvement on all relevant platforms.

TESTED:
CLOCK_MONOTONIC now works and gives plausible results on all
platforms.