Makefile: Exclude empty objects from static library.
Some (or even many) object files are logically empty in a given build.
These are now excluded from linking the static library, which not only
makes the result a bit smaller (while equivalent), but also avoids the
"no symbols" warnings when building it.

Due to the kludgy way fdopendir.o is constructed, this method doesn't
filter it out when it's logically empty (10.10+).  We ignore that for
now, since it simply means that the old behavior remains for that one
object, and we may improve its creation in the future.

TESTED:
Empty objects are filtered as intended (except as noted above), and
all statically-linked tests (added in a subsequent commit) pass.