Rewrite symbol aliases test.
The old symbol aliases test relied on looking up symbols with dlsym(),
which doesn't work with a statically-linked library.  This rewrites it
to use ordinary external references, which will fail if the referenced
symbols are missing.

This means that any errors are build-time errors rather than runtime
errors.  As noted in the comment, there's theoretically a way to avoid
that by using weak references, but we don't bother to wrestle with the
added complications of that approach.

TESTED:
This version passes with both static and dynamic libraries.