Add tests for alternate realpath() versions.
This ensures that all three versions of realpath() (two in 64-bit
builds) have test coverage.  The test for the impossible 64-bit
non-POSIX case is a dummy.

The 10.15 non-POSIX realpath() has a bug (unrelated to
legacy-support) causing the relevant test to fail.  A subsequent
commit excludes that case from the test.

This test uncovered real bugs in the former wrapper implementation.

TESTED:
Without the realpath() fix, fails on 10.5 and on 10.4 with later SDK.
With the fix, passes on all platforms (except as noted above).