Makefile: Add syslib and 'all' test targets.
The syslib targets run versions of the tests linked against the
"system" version of the library, substituting for the normal OS system
library.  This provides a check that the syslib was built correctly.

Targets for all syslib tests and all versions of all tests are added.

TESTED:
Tests build and pass on all platforms.
New targets run all expected tests.