Makefile: Fix superfluous library rebuilds.
The fix for parallel builds made directories separate targets, but
directory dependencies need to be order-only, since their mtimes are
updated whenever files are added, removed, or renamed.  Failure to do
this resulted in intermittent unnecessary rebuilds of libraries when
building tests.  When building directly in the repo, this is a
harmless waste of time, but when building in the MacPorts context,
such rebuilds fail due to ownership and permission issues.  This
change makes directories order-only dependencies, thereby ignoring
their mtimes.

This also removes unnecessary directory dependencies from the tests
that don't use the library.

TESTED:
Ran tests on all platforms.  Intermittent test failures are now gone.