Fix build on 10.4 x86_64
The special treatment for 10.4 in splitandfilterandmergemultiarch was
not being applied appropriately to x86_64, causing a build failure
for fdopendir.

Fixing this bug prior to fixing the realpath() wrapper resulted in
successfully building code that didn't work correctly, though the
realpath() bug wasn't limited to this one configuration.

Note that at the time of this writing, when this case is built as a
port, the architecture is coerced to i386.  Until that is fixed, this
fix won't affect the actual port.

TESTED:
With the realpath() fix, builds successfully on all platforms,
including 10.4 x86_64.