sysconf: Speed up _SC_PHYS_PAGES.
It's a silly waste of time to use the string-lookup version of
sysconf() rather than using the numeric version, even though that's
what Apple's code does.

Also adds comment explaining 32-bit issues.

Also, since the "i386" bug in Apple's code is really a 32-bit bug, the
feature-flag condition is adjusted to be more honest.  This has no
actual effect, since i386 is the only supported 32-bit processor in
the relevant OS versions.  This removes the only reference to
__MPLS_APPLE_I386__, but the definition is retained in case it's
useful in the future.

TESTED:
Passes tests on all platforms.