test_fdopendir: Add fstatat() buffer overrun defense.
At the time of this writing, fstatat() may choose the wrong variant of
stat to return, possibly overflowing the provided buffer.  To defend
against this, we make the buffer the larger of the two possible sizes,
and also move it off the stack for less verbosity when debugging.  See
the comment in the code.

TESTED:
All versions pass, including previously crashing ino32 cases.