Makefile: Use system tools.
Referencing the "bare" ar, ld, lipo, and install_name_tool
opportunistically uses the MacPorts cctools and ld64 versions, without
explicit dependencies.  This adds the explicit /usr/bin/ prefix to all
but lipo to avoid that.  These are still overridable defaults.

Since LD and AR are standard make defaults, defaulting them doesn't
work.  We switch to LDX and ARX to avoid that.

The lipo reference is left as is (since the system lipo is
inadequate), but is expected to go away in the near future.

TESTED:
Builds (including tests) on all platforms, with all applicable SDKs.