Add header support for stat64 with the 10.4 SDK.
This provides support in the headers for the stat64 struct with the
10.4 SDK, as follows:

MacportsLegacySupport.h adds feature flags for stat64 support on 10.4.

The internal __DARWIN_STRUCT_STAT64 macro is defined.

Symbol versioning macros related to INODE64 are added.

The "struct stat64" definition is provided when appropriate, along
with prototypes for the *stat64() functions.

Accommodations are made for the switch from _APPLE_C_SOURCE to
_DARWIN_C_SOURCE between the 10.4 and 10.5 SDKs.

It does not include library support for *stat64() on 10.4, which
is left for a separate commit.

TESTED:
Builds and passes tests	on all platforms.