sys/cdefs: Narrow scope of __has_cpp_attribute fix.
The workaround for the bug in __has_cpp_attribute() is acceptable for
all versions of both clang and GCC, but causes unsuppressible warnings
in affected versions of GCC.  To minimize this, we narrow the scope of
the workaround in the GCC case to just the cases that actually require
it.

TESTED:
Eliminates warnings where possible.
Tests pass on all platforms.