Rename common-priv.h to atcalls.h.
Since the sole purpose of the internal common-priv.h was to support
atcalls, naming it accordingly is more appropriate.  If another
circumstance arises calling for an internal header, it would make more
sense to make it a separate header with an appropriate name.

Also does some whitespace and include-order cleanup.

TESTED:
Tests pass on all platforms.