Try to fix st?ncpy_chk warnings
These tests are intentionally performing illegal operations to verify
that the checks work as expected.  We try to suppress related warnings
as much as possible.

With these changes, all tested versions of clang are fully
well-behaved, in that they emit no warnings and apply the checks as
expected.

Gcc is another story, where some versions fail to suppress the
warnings, and some versions actually get false errors.  The current
state seems to be the best possible under the circumstances.

TESTED:
Behavior across compiler versions is as noted above.
Tests pass on all platforms.