Replace copyfile wrapper with 10.6 copyfile.
After applying many fixes to the old copyfile wrapper, it was still
hopeless on 10.4 due to a fatal internal bug in Apple's code.  This
change throws all that out and imports the entire copyfile
implementation from 10.6.  This should provide the full 10.6 copyfile
on 10.5, and a mostly full implementation on 10.4, after additional
10.4 fixes in the next commit.

This retains the feature of Apple's copyfile that allows it to be
built as a standalone program for easier testing.

TESTED:
Builds and passes the test on 10.5+.
Doesn't yet build on 10.4 (fixed in subsequent commit).