copyfile: Fix build issues on 10.4.
Since the O_SYMLINK definition is missing from the 10.4 fcntl.h, we
provide a default here.  It's the correct value, but whether or not it
actually works correctly in 10.4 is unknown.

Since 10.4 lacks quarantine support, we provide dummy macros for the
quarantine definitions referenced here.

With these changes, 10.4 should now have the full 10.6 copyfile,
except that quarantine-related operations almost certainly don't work,
and special handling of symlinks may not work.

TESTED:
Builds and passes the test on all OS versions, including 10.4.