Add 10.6 copyfile manpage for 10.4 and 10.5.
This is taken from:
https://github.com/apple-oss-distributions/copyfile/blob/copyfile-66.1/copyfile.3

Almost verbatim, but with additions to indicate that QUARANTINE
operations are unavailable on 10.4, and that ACL copying doesn't work
in 10.4 Rosetta.

Adds manpage directory creation to Makefile, including missing case
for the previously added 'which' manpage.

TESTED:
Manpages are installed as appropriate by the relevant targets.