test_copyfile: Avoid broken COPYFILE_ACL on 10.4 Rosetta.
Due to a bug in fstatx_np() in 10.4 Rosetta, COPYFILE_ACL doesn't
work.  In that case, we disable that particular operation to avoid a
test failure.  The remaining three COPYFILE_ALL operations are still
tested.

There's also a bug in 10.4 Rosetta's chmodx_np(), which causes an
internal error when attempting to ensure that the destination is
writable, but that doesn't cause failures in the test case.

TESTED:
Passes on all platforms, including 10.4 Rosetta.