Makefile: Support manual tests that use the library.
Most manual tests don't need to build with the library, and prior to
this change, *no* manual tests (except the old C++ tests) built with
the library.  This adds a naming convention to allow some manual tests
to use the library, currently just the clock_settime() test.

TESTED:
Manual tests with and without the library build as expected.