Make stdio.h include sys/stdio.h for renameat().
The declaration of renameat() is in sys/stdio.h, to match Apple's
arrangement (when it exists) in 10.10+, but Apple also includes
(directly or indirectly) sys/stdio.h from stdio.h, so we need to
duplicate that behavior to ensure that stdio.h provides renameat() as
expected.

Closes: https://trac.macports.org/ticket/71842

TESTED:
Passes all tests, including updated renameat test.