Provide pthread_[f]chdir_np() functions for 10.5+.
This provides both missing prototypes and missing function wrappers
for the pthread_[f]chdir_np() functions, which were introduced as
syscalls in 10.5, but not provided as functions until 10.12, and
not provided as prototypes at all in any SDK.

At present, no SDK for any OS version provides prototypes for these
functions, even when the functions themselves are included in
libSystem.  For that reason, we condition the addition of the
prototypes on the flag _MACPORTS_LEGACY_PTHREAD_CHDIR.  This is
ignored in the 10.4 SDK, since the functions themselves are not
available in 10.4.

The missing function wrappers are provided for 10.5-10.11.  For 10.4,
dummy functions that always fail are provided, though no SDK
declarations are present in the 10.4 SDK, even with the enabling flag.

Closes: https://trac.macports.org/ticket/71265

TESTED:
Passes all tests on all platforms, including the new test.