Rewrite clock_gettime_nsec_np()
This rewrites clock_gettime_nsec_np() as a separate function, rather
than as a wrapper around clock_gettime().  It does not redo the
mach_time scaling, which is left for another commit.  It does
provide some groundwork for improvements to clock_gettime().

Also adds the missing "sandwich" to get a consistent
boottime/timeofday pair, to avoid a race.

TESTED:
Passes new clock test on all platforms, with temporary forgiveness for
inaccurate mach_time scaling on PPC.