Rewrite clock_gettime() and clock_getres().
This rewrites clock_gettime() to be cleaner and to use the support
functions added with the clock_gettime_nsec_np() rewrite.  It
similarly rewrites clock_getres(), including reporting the correct
value for the mach_time-derived clocks.  It does not fix the
low-accuracy mach_time scaling.

TESTED:
Passes new clock test on all platforms, with temporary forgiveness for
inaccurate mach_time scaling on PPC.