Narrow CMSG_DATA fix.
It was originally assumed that the 64-bit bug in the 10.5 CMSG_DATA
definition might apply to 10.4 64-bit as well (even though the code
isn't identical).  It wasn't practical to try the real failing case
(ntpsec's ntpd), so assuming that 10.4 also needed the fix was
conservative.  But a new ntpsec-related test program was able to
confirm that the bug doesn't exist in 64-bit 10.4, so we narrow the
applicability of the fix.

TESTED:
Test still passes on all platforms.