libtest_settime: Report any adjustments to boottime.
Since settimeofday() in 10.12+ adjusts boottime as a kludge to support
CLOCK_MONOTONIC, it's useful to report such changes, and hence which
OS versions actually do it.

To be certain whether lack of adjustments is due to the limited
resolution of boottime in the older OS versions, this mode includes
arranging for the step adjustment to straddle either the second
boundary (for truncation) or the odd half-second boundary (for
rounding).

TESTED:
Demonstrates adjustments in 10.12+, and lack of adjustments in <10.12,
even when timing is arranged to expect an adjustment when
implemented.