Add allheaders tests.
This adds a test which includes all headers which we provide wrappers
for (with a few specific exceptions), as a way to test the headers for
compile problems.  It also includes a script to generate the "master
header", to be used from time to time when the list may have changed.

Identical C++ versions are also provided as manual tests, since C++ is
only usable in more limited circumstances.

NOTE: At the time of this writing, the C++ build of cmath header fails
on 10.4 x86_64, due to an apparent issue with system headers.  10.4
ppc and 10.4 i386 both work.

TESTED:
Builds all headers as expected, including discovering problems before
they were fixed.