time.c: Implement sleep offset.
This provides the sleep offset for the "continuous" time functions
and derivatives.  This only accounts for "static" sleep offsets,
not "dynamic" sleep offsets.  See the comments for more details.

Closes: https://trac.macports.org/ticket/68007

TESTED:
Gives plausible no-sleep results on all platforms.  Sleeps aren't
available in VMs, limiting the test cases for actual sleeps.
Tested with sleeps on 10.4-10.5 ppc and 10.7-10.11 x86_64.