Add branch hints to one-time setups.
This arranges for various "first time" setups to assume that the
already set up case is the normal path, provided that the compiler
supports that feature.

In practice, this seems to have no effect in these particular cases,
though it doesn't hurt.  Apple's code does something similar in a
couple of cases, which is what prompted these additions.  It's not
known whether the compiler is actually ignoring the hints, or whether
the code was already considered optimal without the hints and hence
unaffected.

TESTED:
Builds and passes tests.