Add degenerate *clonefile*() support.
This is a degenerate implementation which only returns errors, since
the real clonefile functions are only valid on APFS, which isn't
supported by any OS that needs this.

It also adds the definition for VOL_CAP_INT_CLONE to sys/attr.h, since
relevant code might be checking that attribute.

Closes: https://trac.macports.org/ticket/72272

TESTED:
Passes tests on all platforms.