stpncpy_chk: Fix build for mismatched SDK.
The __chk_fail() function is only available in the 10.5 and 10.6
libSystem, so building for those OS versions with a 10.7+ SDK gets
an undefined symbol.  This changes it to use dlsym() to look up
the symbol at runtime (which only happens in the failing case).
Weak linking doesn't work without extra hackery.

TESTED:
Now builds with all compatible SDKs, and tests pass.