Fix manual st?ncpy_chk tests.
The previous move to the manual test category neglected to include
the libtest_ prefix to indicate that they needed to be built with
the library, and the test of the change wasn't run on a sufficiently
old OS to notice.

TESTED:
Tests now run on all platforms.  This time for sure.