test_clocks: Add boottime checking and reporting.
This checks the boottime for obviously invalid values (later than the
current time), and optionally reports the value.  This value is used
in computing the sleep offset.

For now, we ignore any boottime error for test failure purposes, since
the actual clock code has a workaround, and we don't want to induce
unnecessary test failures.  This will be changed to be treated as a
test failure after the boottime bug is fixed.

TESTED:
Passes on all platforms.
Includes expected report.