test_stat: Various fixes.
Now verifies that all tv_nsec values are within range.  In reality,
tv_nsec values should always be zero in 10.4, but we don't assume
that; we simply verify that the values are in the allowable range (and
matching where expected as before).  aAlthough atimes can't be compared
in this test, it's still reasonable to check nanosecond validities.

Now uses stack-based stat buffers.  The code was originally written to
use static buffers for the test data for two reasons:

1) The effects of buffer overflow are less nasty.

2) It avoided the need to pass pointers through multiple function
   levels.

But since the buffers are already designed to avoid known
buffer-overflow bugs, and the buffer-overflow bugs have been fixed, #1
is no longer an issue.  Meanwhile, since 64-bit builds under 10.4
don't avoid the bottom 4GiB, the static buffers fail to properly test
64-bit addressing.  So we switch to stack-based buffers, while setting
up static pointers to address #2 without significant code changes.

Now includes test of stat() while open.  It was suspected that the
10.4 ppc64 bug first observed in fstat() and later in stat() might be
affected by whether the file is open, rather than by which call is
being used.  This suspicion turned out to be false, but we keep the
additional test, anyway.

Also captures fileno in fstat() test, to assist debugging.

Also fixes a shadowing warning.

TESTED:
Fails as expected on 10.4 ppc64; otherwise passes on all platforms.
Passes on all platforms with the 10.4 ppc64 fix applied.