test_attrlist: New consolidated attrlist test.
This test covers the traditional [gs]etattrlist() functions, as well
as the newer f[gs]etattrlist() functions.  It replaces the old
f[gs]etattrlist tests.

It adds "null" get/set test cases, to test what was believed to be
a necessary workaround for a bug in such cases, though the tests
pass without that workaround.

TESTED:
Without 10.4 ppc64 fix, fails on 10.4 ppc64 and passes on others.
With fix, passes on all platforms.