fxattrlist: New combined source for f[gs]attrlist.
This combines the similar functions into one source, to permit some
code sharing.  There were originally two types of code potentially
common to both, but one is now handled in a different way, and the
other has proved to be unnecessary.  Nevertheless, it's a better code
organization in general.

The previous implementations included special handling for the "null"
case, but after adding tests for those cases, there is no evidence
that such handling was actually necessary.

TESTED:
Tests pass on all platforms.