test_stat: Rework open/close logic.
This reworks the handling of opening and closing the test file or
directory, including reporting these events in verbose mode.  It was
originally thought that the open/closed status might affect the 10.4
ppc64 bug reporting garbage tv_nsec values, but this turned out not to
be the case.

Also fixes a recurring typo in many link-related messages.

TESTED:
Passes on all platforms, except 10.4 ppc64 when fix is omitted.
Verbose output is as expected.