Add tool to get some stat() info.
This exists primarily because the normal tools don't report subsecond
timestamps, which are an issue in a 10.4 ppc64 bug.  Although subecond
timestamps are largely unsupported, in such cases they should be zero,
not garbage as in 10.4 ppc64.  This tool helps illustrate the issue.

TESTED:
Works as intended on all platforms.